refactor logger and packed source to seperate files
