+++ utility/gdre_packed_source.cpp
@@ -1,0 +1,139 @@
+#include "gdre_packed_source.h"
+#include "core/io/file_access_encrypted.h"
+#include "gdre_settings.h"
+
+bool GDREPackedSource::try_open_pack(const String &p_path, bool p_replace_files, uint64_t p_offset) {
+	if (p_path.get_extension().to_lower() == "apk") {
+		return false;
+	}
+	String pck_path = p_path.replace("_GDRE_a_really_dumb_hack", "");
+	Ref<FileAccess> f = FileAccess::open(pck_path, FileAccess::READ);
+	if (f.is_null()) {
+		return false;
+	}
+
+	f->seek(p_offset);
+
+	uint32_t magic = f->get_32();
+
+	if (magic != PACK_HEADER_MAGIC) {
+		// loading with offset feature not supported for self contained exe files
+		if (p_offset != 0) {
+			ERR_FAIL_V_MSG(false, "Loading self-contained executable with offset not supported.");
+		}
+
+		//maybe at the end.... self contained exe
+		f->seek_end();
+		f->seek(f->get_position() - 4);
+		magic = f->get_32();
+		if (magic != PACK_HEADER_MAGIC) {
+			return false;
+		}
+		f->seek(f->get_position() - 12);
+
+		uint64_t ds = f->get_64();
+		f->seek(f->get_position() - ds - 8);
+
+		magic = f->get_32();
+		if (magic != PACK_HEADER_MAGIC) {
+			return false;
+		}
+	}
+
+	uint32_t version = f->get_32();
+	uint32_t ver_major = f->get_32();
+	uint32_t ver_minor = f->get_32();
+	uint32_t ver_rev = f->get_32(); // patch number, not used for validation.
+
+	if (version > PACK_FORMAT_VERSION) {
+		ERR_FAIL_V_MSG(false, "Pack version unsupported: " + itos(version) + ".");
+	}
+
+	uint32_t pack_flags = 0;
+	uint64_t file_base = 0;
+
+	if (version == 2) {
+		pack_flags = f->get_32();
+		file_base = f->get_64();
+	}
+
+	bool enc_directory = (pack_flags & PACK_DIR_ENCRYPTED);
+
+	for (int i = 0; i < 16; i++) {
+		//reserved
+		f->get_32();
+	}
+
+	uint32_t file_count = f->get_32();
+
+	if (enc_directory) {
+		Ref<FileAccessEncrypted> fae = memnew(FileAccessEncrypted);
+		if (fae.is_null()) {
+			GDRESettings::get_singleton()->_set_error_encryption(true);
+			ERR_FAIL_V_MSG(false, "Can't open encrypted pack directory.");
+		}
+
+		Vector<uint8_t> key;
+		key.resize(32);
+		for (int i = 0; i < key.size(); i++) {
+			key.write[i] = script_encryption_key[i];
+		}
+
+		Error err = fae->open_and_parse(f, key, FileAccessEncrypted::MODE_READ, false);
+		if (err) {
+			GDRESettings::get_singleton()->_set_error_encryption(true);
+			ERR_FAIL_V_MSG(false, "Can't open encrypted pack directory.");
+		}
+		f = fae;
+	}
+
+	// Everything worked, now set the data
+	Ref<GDRESettings::PackInfo> pckinfo;
+	pckinfo.instantiate();
+	pckinfo->init(
+			pck_path, ver_major, ver_minor, ver_rev, version, pack_flags, file_base, file_count,
+			itos(ver_major) + "." + itos(ver_minor) + "." + itos(ver_rev), GDRESettings::PackInfo::PCK);
+	GDRESettings::get_singleton()->add_pack_info(pckinfo);
+
+	for (uint32_t i = 0; i < file_count; i++) {
+		uint32_t sl = f->get_32();
+		CharString cs;
+		cs.resize(sl + 1);
+		f->get_buffer((uint8_t *)cs.ptr(), sl);
+		cs[sl] = 0;
+
+		String path;
+		path.parse_utf8(cs.ptr());
+
+		ERR_FAIL_COND_V_MSG(path.get_file().find("gdre_") != -1, false, "Tried to load a gdre file?!?!");
+
+		uint64_t ofs = file_base + f->get_64();
+		uint64_t size = f->get_64();
+		uint8_t md5[16];
+		uint32_t flags = 0;
+		f->get_buffer(md5, 16);
+		if (version == 2) {
+			flags = f->get_32();
+		}
+		// add the file info to settings
+		PackedData::PackedFile pf;
+		pf.offset = ofs + p_offset;
+		memcpy(pf.md5, md5, 16);
+		pf.size = size;
+		pf.encrypted = flags & PACK_FILE_ENCRYPTED;
+		pf.pack = p_path;
+		pf.src = this;
+		Ref<PackedFileInfo> pf_info;
+		pf_info.instantiate();
+		pf_info->init(path, &pf);
+		GDRESettings::get_singleton()->add_pack_file(pf_info);
+		// use the corrected path, not the raw path
+		path = pf_info->get_path();
+		PackedData::get_singleton()->add_path(pck_path, path, ofs + p_offset, size, md5, this, p_replace_files, (flags & PACK_FILE_ENCRYPTED));
+	}
+
+	return true;
+}
+Ref<FileAccess> GDREPackedSource::get_file(const String &p_path, PackedData::PackedFile *p_file) {
+	return memnew(FileAccessPack(p_path, *p_file));
+}
